+++ src/license/licenseInfos/CCBY25LicenseInfos.ts
