+++ src/license/licenseInfos/LGPLLicenseInfos.ts
