+++ src/license/licenseInfos/LicenseInfos.ts
